rename: var name and files
